Set up 433 for duplex transmissions. Not sending full data. Only sending movement commands to the rover. Only returning system time to base.
The getRF() and putRF() functions are different for the rover and the base.
Rover will block indefinitely on receive and will not check cts on send.
Base will wait until buffer is empty to send mobility but will only
briefly wait to see if there is any data in the receive buffer before
timing out.